'which' is deprecated in Debian sid
Using the `which` binary on a Debian sid installation today emits
deprecation warnings that suggest using `command -v` instead.

Unfortunately I found that on my MacOS install I found that
/usr/bin/command is a shell script:

builtin `echo ${0##*/} | tr \[:upper:] \[:lower:]` ${1+"$@"}

as well as a shell builtin:
$ command -v command
command

...which is not helpful from Ruby.

I did some research and it seems there is no canonical implementation of
Which in ruby standard lib proper, but there is ptools that adds which
to the File class.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>